nu_system::ForegroundChild::wait(): Return nu_system::ExitStatus
This will allow us to use nix::sys::wait::waitpid() in this function,
which doesn't give us access to the numeric wait status that we need to
construct std::process::ExitStatus.

We want to be able to use nix::sys::wait, because std::process::Child
doesn't allow us to wait for a child to be stopped (WUNTRACED option to
waitpid(2)).