Support lists in `build_query`
Swagger supports lists (a.k.a arrays) in query parameters:

  https://swagger.io/docs/specification/v3_0/serialization/

It supports three different styles:

  - explode=true
  - spaceDelimited
  - pipeDelimited

With explode=true being the default and hence most common. It is also
the hardest to use inside nushell, as the others are just a `string
join` away. This commit adds lists with the explode=true format.