Clearer build_query list example
"the a[] may be a bit confusing here. If you read it it might imply you
need to append it to be able to use the list form but it will work with
any key. (even if the consuming API may have some ideas about what it
permits there)"

Thanks: shoulderbach